feat: 修复网络协议和添加客户端测试工具
- 修复 EncodePacket 包格式从 4 字节头改为 6 字节头
- 修复 processBuffer 解析 6 字节头
- 修复 handleLogin/handleNewAccount 数据偏移从 data[2:] 改为 data[14:]
- 修复 handleSelectServer 服务器索引从 data[2] 改为 data[14:]
- 修复 sendServerInfo 使用全局配置而不是默认配置
- 重写 RunGate 为每个客户端建立独立的 M2Server 连接
- 添加 cmd/client_test 完整流程测试客户端
- 更新 README.md 添加客户端测试工具说明

+++ cmd/testclient/main.go
@@ -180,7 +180,7 @@ func (c *Client) testMovement() {
 	
 	for _, t := range tests {
 		fmt.Printf("Sending %s (%d)...\n", t.name, t.ident)
-		c.sendMessage(t.ident, t.param, t.tag, 0, 0)
+		c.sendMessage(t.ident, int32(t.param), t.tag, 0, 0)
 		time.Sleep(50 * time.Millisecond)
 	}
 }
